pyproject.toml
[tool.poetry]
name = "pyhtmlcv"
version = "1.0.0"
description = "Python-based HTML CV generator"
authors = ["Janne Enberg <janne.enberg@lietu.net>"]
license = "BSD-3-clause"

[tool.poetry.dependencies]
python = "^3.7"
Jinja2 = "2.10.1"
libsass = "0.18.0"
pathlib2 = "2.3.3"
six = "1.12.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
